+++ app/controllers/api_controller.rb
@@ -38,7 +38,7 @@ class ApiController < ActionController::API
   protected
 
   def conflict_error?(resource)
-    resource.errors[:updated_in_db_at].include?(ApplicationRecord::CONFLICT_ERROR_MESSAGE)
+    resource.errors[:updated_in_db_at].any? { |error| error.include?(ApplicationRecord::CONFLICT_ERROR_MESSAGE) }
   end
 
   def current_account_list